scripts/requirements.txt
openai
PyYAML
python-dotenv